Base for weather APi implemented on Weather.jsx
If crash there/no data viewing, it might be because the API is a bit unstable (as read on API site comments)

Try reloading the API-page to see if it works then. If it keeps malfunctioning, cut down the weather data to just one Endpoint (Current?)
Since having to many request to the API could make it freeze/mix all the different GET:s into one same for all endpoints. Making the pressure less on the server could get the pressure down on it.